sentiment_analysis/code/requirements.txt
streamlit==1.29.0
python==3.11.6
nltk==3.8.1
pandas==2.1.3
seaborn==0.13.0
matplotlib-base==3.8.2
regex==2023.10.3
boto3==1.33.11  